// src/models/landing-page-product-data.ts
export interface LandingPageProductData {
  publication_title: string;
  publication_url: string;
  product_sku: string;
  product_price: string;
  product_category: string;
  product_city: string;
  product_zone: string;
  product_neighborhood: string;
  business_type: string;
  product_type: string;
}